+++ Heading.mjs
@@ -1,5 +1,4 @@
 import classNameProp from "class-name-prop";
-import PropTypes from "prop-types";
 import React from "react";
 import LinkText from "./LinkText.mjs";
 
@@ -25,16 +24,7 @@ const Heading = React.forwardRef(
     )
 );
 
-if (typeof process === "object" && process.env.NODE_ENV !== "production") {
+if (typeof process === "object" && process.env.NODE_ENV !== "production")
   Heading.displayName = "Heading";
-  Heading.propTypes = {
-    level: PropTypes.oneOf([1, 2, 3, 4, 5, 6]),
-    size: PropTypes.oneOf([1, 2, 3, 4, 5, 6]),
-    id: PropTypes.string,
-    className: PropTypes.string,
-    style: PropTypes.object,
-    children: PropTypes.node.isRequired,
-  };
-}
 
 export default Heading;