+++ LinkActive.mjs
@@ -1,6 +1,5 @@
 import NextLink from "next/link.js";
 import NextRouter from "next/router.js";
-import PropTypes from "prop-types";
 import React from "react";
 
 const LinkActive = ({ children, ...props }) => {
@@ -13,11 +12,5 @@ const LinkActive = ({ children, ...props }) => {
     })
   );
 };
-
-if (typeof process === "object" && process.env.NODE_ENV !== "production")
-  LinkActive.propTypes = {
-    href: PropTypes.string,
-    children: PropTypes.element.isRequired,
-  };
 
 export default LinkActive;