+++ LinkCard.mjs
@@ -1,7 +1,5 @@
 import classNameProp from "class-name-prop";
-import PropTypes from "prop-types";
 import React from "react";
-import propTypeChildren from "./propTypeChildren.mjs";
 
 const LinkCard = React.forwardRef(({ active, className, ...props }, ref) =>
   React.createElement("a", {
@@ -15,13 +13,7 @@ const LinkCard = React.forwardRef(({ active, className, ...props }, ref) =>
   })
 );
 
-if (typeof process === "object" && process.env.NODE_ENV !== "production") {
+if (typeof process === "object" && process.env.NODE_ENV !== "production")
   LinkCard.displayName = "LinkCard";
-  LinkCard.propTypes = {
-    active: PropTypes.bool,
-    className: PropTypes.string,
-    children: propTypeChildren.isRequired,
-  };
-}
 
 export default LinkCard;