+++ Nav.mjs
@@ -1,7 +1,5 @@
 import classNameProp from "class-name-prop";
-import PropTypes from "prop-types";
 import React from "react";
-import propTypeChildren from "./propTypeChildren.mjs";
 
 const Nav = React.forwardRef(({ className, ...props }, ref) =>
   React.createElement("nav", {
@@ -11,12 +9,7 @@ const Nav = React.forwardRef(({ className, ...props }, ref) =>
   })
 );
 
-if (typeof process === "object" && process.env.NODE_ENV !== "production") {
+if (typeof process === "object" && process.env.NODE_ENV !== "production")
   Nav.displayName = "Nav";
-  Nav.propTypes = {
-    className: PropTypes.string,
-    children: propTypeChildren.isRequired,
-  };
-}
 
 export default Nav;