+++ Picture.mjs
@@ -1,7 +1,5 @@
 import classNameProp from "class-name-prop";
-import PropTypes from "prop-types";
 import React from "react";
-import propTypeChildren from "./propTypeChildren.mjs";
 
 const Picture = React.forwardRef(
   ({ width, height, className, style, ...props }, ref) =>
@@ -17,15 +15,7 @@ const Picture = React.forwardRef(
     })
 );
 
-if (typeof process === "object" && process.env.NODE_ENV !== "production") {
+if (typeof process === "object" && process.env.NODE_ENV !== "production")
   Picture.displayName = "Picture";
-  Picture.propTypes = {
-    width: PropTypes.number.isRequired,
-    height: PropTypes.number.isRequired,
-    className: PropTypes.string,
-    style: PropTypes.object,
-    children: propTypeChildren.isRequired,
-  };
-}
 
 export default Picture;