+++ Scroll.mjs
@@ -1,5 +1,4 @@
 import classNameProp from "class-name-prop";
-import PropTypes from "prop-types";
 import React from "react";
 
 const Scroll = React.forwardRef(({ className, ...props }, ref) =>
@@ -10,12 +9,7 @@ const Scroll = React.forwardRef(({ className, ...props }, ref) =>
   })
 );
 
-if (typeof process === "object" && process.env.NODE_ENV !== "production") {
+if (typeof process === "object" && process.env.NODE_ENV !== "production")
   Scroll.displayName = "Scroll";
-  Scroll.propTypes = {
-    className: PropTypes.string,
-    children: PropTypes.node.isRequired,
-  };
-}
 
 export default Scroll;