+++ Select.mjs
@@ -1,7 +1,5 @@
 import classNameProp from "class-name-prop";
-import PropTypes from "prop-types";
 import React from "react";
-import propTypeChildren from "./propTypeChildren.mjs";
 import useCustomValidity from "./useCustomValidity.mjs";
 import useMergedRef from "./useMergedRef.mjs";
 import useOnFocusReportValidity from "./useOnFocusReportValidity.mjs";
@@ -21,13 +19,7 @@ const Select = React.forwardRef(
   }
 );
 
-if (typeof process === "object" && process.env.NODE_ENV !== "production") {
+if (typeof process === "object" && process.env.NODE_ENV !== "production")
   Select.displayName = "Select";
-  Select.propTypes = {
-    validationMessage: PropTypes.string,
-    className: PropTypes.string,
-    children: propTypeChildren.isRequired,
-  };
-}
 
 export default Select;