+++ changelog.md
@@ -11,6 +11,9 @@
 - Removed the package main index module; deep imports must be used.
 - Shortened public module deep import paths, removing the `/public/`.
 - The API is now ESM in `.mjs` files instead of CJS in `.js` files, [accessible via `import` but not `require`](https://nodejs.org/dist/latest/docs/api/esm.html#require).
+- Removed the [`prop-types`](https://npm.im/prop-types) dependency and related functionality:
+  - Removed the `propTypeChildren` package export.
+  - Removed the `propTypes` property from React components.
 - More modern JavaScript syntax.
 - Removed some CSS vendor prefixes.
 - Switched back to using `React.createElement` instead of the [the new React JSX runtime](https://reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html).